+++ planar_magnetics/windings/windings.py
@@ -4,6 +4,24 @@ from pathlib import Path
 
 
 class Winding:
+    def mirror_x(self):
+        """Mirror the winding about the x-axis"""
+        mirrored = self
+        mirrored.polygon = self.polygon.mirror_x()
+        return mirrored
+
+    def mirror_y(self):
+        """Mirror the winding about the y-axis"""
+        mirrored = self
+        mirrored.polygon = self.polygon.mirror_y()
+        return mirrored
+
+    def rotate(self, angle: float):
+        """Rotate a winding about its center"""
+        rotated = self
+        rotated.polygon = self.polygon.rotate_about(self.at, angle)
+        return rotated
+
     def plot(self, ax=None, max_angle: float = math.pi / 36):
         self.polygon.plot(ax, max_angle)